[FP-77] Fix: 토큰 인터셉터 수정
